hv: ioapic: Change interrupt pin to uint32_t
Change the type from uint8_t to uint32_t for interrupt pin

Tracked-On: #861
Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/include/arch/x86/ioapic.h
@@ -14,14 +14,14 @@
 void ioapic_setup_irqs(void);
 
 bool ioapic_irq_is_gsi(uint32_t irq);
-uint8_t ioapic_irq_to_pin(uint32_t irq);
+uint32_t ioapic_irq_to_pin(uint32_t irq);
 
 /**
  * @brief Get irq num from pin num
  *
  * @param[in]	pin The pin number
  */
-uint32_t ioapic_pin_to_irq(uint8_t pin);
+uint32_t ioapic_pin_to_irq(uint32_t pin);
 
 /**
  * @brief Set the redirection table entry
@@ -51,17 +51,17 @@ void resume_ioapic(void);
 void ioapic_gsi_mask_irq(uint32_t irq);
 void ioapic_gsi_unmask_irq(uint32_t irq);
 
-void ioapic_get_rte_entry(void *ioapic_addr, uint8_t pin, union ioapic_rte *rte);
+void ioapic_get_rte_entry(void *ioapic_addr, uint32_t pin, union ioapic_rte *rte);
 
 struct gsi_table {
 	uint8_t ioapic_id;
-	uint8_t pin;
+	uint32_t pin;
 	void  *addr;
 };
 
 void *ioapic_get_gsi_irq_addr(uint32_t irq_num);
 uint32_t ioapic_get_nr_gsi(void);
-uint8_t get_pic_pin_from_ioapic_pin(uint8_t pin_index);
-bool ioapic_is_pin_valid(uint8_t pin);
+uint32_t get_pic_pin_from_ioapic_pin(uint32_t pin_index);
+bool ioapic_is_pin_valid(uint32_t pin);
 
 #endif /* IOAPIC_H */
